Cleaned up project files [SDK-1564] (#378)
* Clean up project files

* Update README.md
