+++ Auth0.podspec
@@ -1,22 +1,6 @@
 version = `agvtool mvers -terse1`.strip
 
 web_auth_files = [
-  'Auth0/_ObjectiveWebAuth.swift',
-  'Auth0/ControllerModalPresenter.swift',
-  'Auth0/OAuth2Grant.swift',
-  'Auth0/AuthTransaction.swift',
-  'Auth0/TransactionStore.swift',
-  'Auth0/WebAuth.swift',
-  'Auth0/WebAuthError.swift',
-  'Auth0/SafariWebAuth.swift',
-  'Auth0/AuthSession.swift',
-  'Auth0/SafariSession.swift',
-  'Auth0/SafariAuthenticationSession.swift',
-  'Auth0/SafariAuthenticationCallback.swift',
-  'Auth0/SilentSafariViewController.swift',
-  'Auth0/NativeAuth.swift',
-  'Auth0/AuthProvider.swift',
-  'Auth0/BioAuthentication.swift',
   'Auth0/A0ChallengeGenerator.h',
   'Auth0/A0ChallengeGenerator.m',
   'Auth0/A0RSA.h',
@@ -25,20 +9,45 @@ web_auth_files = [
   'Auth0/A0SHA.m',
   'Auth0/A0SimpleKeychain+RSAPublicKey.swift',
   'Auth0/Array+Encode.swift',
+  'Auth0/AuthCancelable.swift',
+  'Auth0/AuthProvider.swift',
+  'Auth0/AuthSession.swift',
+  'Auth0/AuthTransaction.swift',
+  'Auth0/AuthenticationServicesSession.swift',
+  'Auth0/AuthenticationServicesSessionCallback.swift',
+  'Auth0/BaseAuthTransaction.swift',
+  'Auth0/BaseWebAuth.swift',
+  'Auth0/BioAuthentication.swift',
   'Auth0/ClaimValidators.swift',
   'Auth0/IDTokenSignatureValidator.swift',
   'Auth0/IDTokenValidator.swift',
   'Auth0/IDTokenValidatorContext.swift',
-  'Auth0/JWT+Header.swift',
   'Auth0/JWK+RSA.swift',
+  'Auth0/JWT+Header.swift',
   'Auth0/JWTAlgorithm.swift',
+  'Auth0/NativeAuth.swift',
+  'Auth0/OAuth2Grant.swift',
+  'Auth0/SessionCallbackTransaction.swift',
+  'Auth0/SessionTransaction.swift',
+  'Auth0/TransactionStore.swift',
+  'Auth0/WebAuth.swift',
+  'Auth0/WebAuthError.swift',
+  'Auth0/_ObjectiveWebAuth.swift'
+]
+
+ios_files = [
+  'Auth0/ControllerModalPresenter.swift',
+  'Auth0/MobileWebAuth.swift',
+  'Auth0/SafariSession.swift',
+  'Auth0/SilentSafariViewController.swift',
   'Auth0/UIApplication+Shared.swift'
 ]
 
 watchos_exclude_files = [
   *web_auth_files,
-  'Auth0/CredentialsManagerError.swift',
-  'Auth0/CredentialsManager.swift'
+  *ios_files,
+  'Auth0/CredentialsManager.swift',
+  'Auth0/CredentialsManagerError.swift'
 ]
 
 Pod::Spec.new do |s|
@@ -66,16 +75,16 @@ Pod::Spec.new do |s|
   s.ios.dependency 'SimpleKeychain'
   s.ios.dependency 'JWTDecode'
   s.osx.source_files = 'Auth0/*.swift'
-  s.osx.exclude_files = web_auth_files
+  s.osx.exclude_files = [*web_auth_files, *ios_files]
   s.osx.dependency 'SimpleKeychain'
   s.osx.dependency 'JWTDecode'
   s.watchos.source_files = 'Auth0/*.swift'
   s.watchos.exclude_files = watchos_exclude_files
   s.watchos.dependency 'JWTDecode'
   s.tvos.source_files = 'Auth0/*.swift'
-  s.tvos.exclude_files = web_auth_files
+  s.tvos.exclude_files = [*web_auth_files, *ios_files]
   s.tvos.dependency 'SimpleKeychain'
   s.tvos.dependency 'JWTDecode'
 
-  s.swift_versions = ['4.0', '4.1', '4.2', '5.0']
+  s.swift_versions = ['4.0', '4.1', '4.2', '5.0', '5.1', '5.2']
 end
