+++ Auth0/ControllerModalPresenter.swift
@@ -53,4 +53,5 @@ struct ControllerModalPresenter {
             return root
         }
     }
+
 }
