+++ Auth0/SessionCallbackTransaction.swift
@@ -1,6 +1,6 @@
-// AuthTransaction.swift
+// SessionCallbackTransaction.swift
 //
-// Copyright (c) 2016 Auth0 (http://auth0.com)
+// Copyright (c) 2020 Auth0 (http://auth0.com)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,18 +20,23 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-/**
-Represents an ongoing Auth transaction with an Identity Provider (Auth0 or a third party).
+class SessionCallbackTransaction: NSObject, AuthTransaction {
 
-The Auth will be done outside of application control, Safari or third party application.
-The only way to communicate the results back is using a url with a registered custom scheme in your application so the OS can open it on success/failure.
-When that happens the OS will call a method in your `AppDelegate` and that is where you need to handle the result.
+    var state: String?
+    var authSession: AuthSession?
+    let callback: (Bool) -> Void
 
-- important: Only one AuthTransaction can be active at a given time for Auth0.swift, if you start a new one before finishing the current one it will be cancelled.
-*/
-public protocol AuthTransaction: AuthResumable, AuthCancelable {
+    init(callback: @escaping (Bool) -> Void) {
+        self.callback = callback
+    }
 
-    /// value of the OAuth 2.0 state parameter. It must be a cryptographically secure randon string used to protect the app with request forgery.
-    var state: String? { get }
+    func cancel() {
+        self.callback(false)
+    }
+
+    func handleUrl(_ url: URL) -> Bool {
+        self.callback(true)
+        return true
+    }
 
 }
